requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
tensorflow-cpu==2.17.1
python-dateutil==2.9.0.post0
